requirements.txt
affine==2.2.2
alembic==1.0.8
argh==0.26.2
arrow==0.13.0
attrs==18.2.0
Babel==2.6.0
bcrypt==3.1.6
blinker==1.4
certifi==2018.11.29
cftime==1.0.3.4
chardet==3.0.4
Click==7.0
click-plugins==1.0.4
cligj==0.5.0
Flask==1.0.2
Flask-BabelEx==0.9.3
Flask-Cors==3.0.7
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-marshmallow==0.9.0
Flask-Migrate==2.4.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
geopyspark==0.4.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.7
MarkupSafe==1.1.0
marshmallow==2.18.1
marshmallow-sqlalchemy==0.16.0
netCDF4==1.4.2
numpy==1.16.0
passlib==1.7.1
pathtools==0.1.2
proj==0.1.0
protobuf==3.6.1
psycopg2==2.7.7
py4j==0.10.8.1
pyparsing==2.3.1
pyspark==2.4.0
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2018.9
PyYAML==5.1
rasterio==1.0.13
requests==2.21.0
Shapely==1.7a1
six==1.12.0
snuggs==1.4.2
speaklater==1.3
SQLAlchemy==1.3.3
urllib3==1.24.2
uWSGI==2.0.18
watchdog==0.9.0
Werkzeug==0.14.1
WTForms==2.2.1
